Add files via upload
Project Two expands on the review and recommendations from Project One to implement mitigation methods for previously identified vulnerabilities. In addition, Artemis Financial requested a file verification function that utilizes checksum in Project Two, which is also included in this application. Notably, the code provided by the course was empty, aside from the pom.xml that identified package versions.